+++ src/flutter_pty.c
@@ -12,3 +12,11 @@
 #include "unix/pty_unix_spawn.c"
 #include "unix/pty_unix_session.c"
 #endif
+
+#if defined(FLUTTER_PTY2_INCLUDE_COMMON_SOURCES)
+#include "common/pty_debug_stats.c"
+#include "common/pty_error.c"
+#include "common/pty_event.c"
+#include "common/pty_refcount.c"
+#include "common/pty_write_queue.c"
+#endif